Move the relevant for zmqActor tests to test_demos
